Regained link acct functionality, file cleanup
