+++ src/App.js
@@ -3,9 +3,9 @@ import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import React, { useState, useEffect, useCallback } from "react";
 
 //Plaid
-import { usePlaidLink } from 'react-plaid-link';
-import './Plaid.scss';
-import './App.css';
+import { usePlaidLink } from "react-plaid-link";
+import "./Plaid.scss";
+import "./App.css";
 
 //Components
 import LogIn from "./components/LogIn";
@@ -18,7 +18,7 @@ import SideNav from "./components/SideNav/SideNav";
 import Trends from "./components/Trends";
 import AccountView from "./AccountComponents/AccountView";
 import AllTransactions from "./components/TransactionsTable/AllTransactions";
-import Balances from "./components/AccountBalances"
+import Balances from "./components/AccountBalances";
 import NotFound from "./components/NotFound";
 
 //Redux
@@ -116,23 +116,22 @@ function App(props) {
   // }
 
   return (
-   
-      <div className='App'>
-        {isLoggedIn ? (
-          <div>
-            {isAdmin ? (
-              <div>
-                {/* <Navbar
+    <div className="App">
+      {isLoggedIn ? (
+        <div>
+          {isAdmin ? (
+            <div>
+              {/* <Navbar
                   open={open}
                   ready={ready}
                   transactions={props.transactions[0]}
                   transactions2={props.transactions[1]}
                 />
                  */}
-              </div>
-            ) : (
-              <div>
-                {/* <div className='navbar-logged-in'>
+            </div>
+          ) : (
+            <div>
+              {/* <div className='navbar-logged-in'>
               
                   <Navbar
                     open={open}
@@ -141,76 +140,79 @@ function App(props) {
                     transactions2={props.transactions[1]}
                   />
                 </div> */}
-               
 
-                <div className="app-container">
-                  <Routes>
-                    <Route
-                      exact
-                      path="/link"
-                      element={
-                        <LinkAccount
-                          open={open}
-                          ready={ready}
-                          transactions={props.transactions[0]}
-                          transactions2={props.transactions[1]}
-                        />
-                      }
-                    />
-                    <Route
-                        path="/*"
-                        element={<Navigate replace to="/dashboard" />}
+              <div className="app-container">
+                <Routes>
+                  <Route
+                    exact
+                    path="/link"
+                    element={
+                      <LinkAccount
+                        open={open}
+                        ready={ready}
+                        // transactions={props.transactions[0]}
+                        // transactions2={props.transactions[1]}
                       />
-                    <Route
-                      exact
-                      path="/dashboard"
-                      element={
-                        <Dashboard transactions={props.transactions[0]} />
-                      }
-                    />
-                    <Route
-                      exact
-                      path="/"
-                      element={<Navigate replace to="/dashboard" />}
-                    />
-                    <Route
-                      exact
-                      path="/login"
-                      element={<Navigate replace to="/dashboard" />}
-                    />
-                    <Route
-                      exact
-                      path="/signup"
-                      element={<Navigate replace to="/dashboard" />}
-                    />
-                    <Route exact path="/budget" element={<BudgetApp />} />
-                    <Route exact path="/trends" element={<Trends />} />
-                    <Route
-                      exact
-                      path="/transactions"
-                      element={<AllTransactions />}
-                    />
-                    <Route exact path="/account" element={<AccountView />} />
-                    <Route exact path="/balances" element={<Balances />} />
-                    <Route exact path="*" element={<NotFound />} />
-                  </Routes>
-                </div>
+                    }
+                  />
+                  <Route
+                    path="/*"
+                    element={<Navigate replace to="/dashboard" />}
+                  />
+                  <Route
+                    exact
+                    path="/dashboard"
+                    element={
+                      <Dashboard
+                        open={open}
+                        ready={ready}
+                        // transactions={props.transactions[0]}
+                        // transactions2={props.transactions[1]}
+                      />
+                    }
+                  />
+                  <Route
+                    exact
+                    path="/"
+                    element={<Navigate replace to="/dashboard" />}
+                  />
+                  <Route
+                    exact
+                    path="/login"
+                    element={<Navigate replace to="/dashboard" />}
+                  />
+                  <Route
+                    exact
+                    path="/signup"
+                    element={<Navigate replace to="/dashboard" />}
+                  />
+                  <Route exact path="/budget" element={<BudgetApp />} />
+                  <Route exact path="/trends" element={<Trends />} />
+                  <Route
+                    exact
+                    path="/transactions"
+                    element={<AllTransactions />}
+                  />
+                  <Route exact path="/account" element={<AccountView />} />
+                  <Route exact path="/balances" element={<Balances />} />
+                  <Route exact path="*" element={<NotFound />} />
+                </Routes>
               </div>
-            )}
-          </div>
-        ) : (
-          <div>
-            {/* for users who aren't logged in */}
-            <Navbar />
-            <Routes>
-              <Route exact path="/signup" element={<SignUp />} />
-              <Route exact path="/login" element={<LogIn />} />
-              <Route exact path="/" element={<LogIn />} />
-            </Routes>
-          </div>
-        )}
-      </div>
-  
+            </div>
+          )}
+        </div>
+      ) : (
+        <div>
+          {/* for users who aren't logged in */}
+          <Navbar />
+          <Routes>
+            <Route exact path="/signup" element={<SignUp />} />
+            <Route exact path="/login" element={<LogIn />} />
+            <Route exact path="/" element={<LogIn />} />
+          </Routes>
+        </div>
+      )}
+    </div>
   );
 }
